# website/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RinaDown API",
    "description": "RinaDown 桌面应用的本机 HTTP API。仅监听 `127.0.0.1`（默认端口 17800，可在 设置 → API 服务 中修改）。\n\n- **takeover**（脚本接管）：油猴脚本提交下载，进入快速下载确认流程；需 `X-RinaDown-Client` 头\n- **aria2**：aria2 JSON-RPC 兼容垫片，「发送到 aria2」类脚本与 AriaNg 可直接对接\n- **management**（管理 API）：任务查询/创建/暂停/恢复/删除与队列查询，供 MCP/自动化客户端使用；强制要求 token（`Authorization: Bearer <token>` 或 `X-RinaDown-Token` 头）",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:17800",
      "description": "本机 API 服务（默认端口）"
    }
  ],
  "paths": {
...
    "/jsonrpc": {
      "post": {
        "tags": [
          "aria2"
        ],
        "summary": "aria2 客户端约定：HTTP 层始终 200，错误在 JSON-RPC envelope 内表达。",
        "operationId": "jsonrpc",
        "responses": {
          "200": {
            "description": "JSON-RPC 响应（错误在 envelope 内表达）。支持方法：aria2.addUri / aria2.getVersion / aria2.getGlobalStat / system.multicall / system.listMethods；token 可经 X-RinaDown-Token 头或 params[0]=\"token:xxx\" 传递"
          }
        },
        "security": [
          {
            "tokenHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "tokenHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RinaDown-Token"
      }
    }
  },
...
}